Fix: RepoRawFile
As of GitLab 6.6, the /projects/:id/repository/blobs/:sha now return only text/plain Content-Type.
Before this there was a bug in the API where gitlab checks the requested filename mime and return a response with the associated Content-Type
Plus, for JSON requested files, the file was double encoded. This now fixed.
See: gitlabhq/gitlabhq#6325